make oam address a byte so it wraps

+++ lib/ppu.go
@@ -87,7 +87,7 @@ type PPUState struct {
 
     /* sprite memory */
     OAM []byte `json:"oam"`
-    OAMAddress int `json:"oamaddress"`
+    OAMAddress byte `json:"oamaddress"`
 
     oamSprites []Sprite
 
@@ -188,11 +188,11 @@ func (ppu *PPUState) SetScreenBMirror(){
 }
 
 func (ppu *PPUState) SetOAMAddress(value byte){
-    ppu.OAMAddress = int(value)
+    ppu.OAMAddress = value
 }
 
 func (ppu *PPUState) WriteOAM(value byte){
-    if ppu.OAMAddress < len(ppu.OAM) {
+    if int(ppu.OAMAddress) < len(ppu.OAM) {
         ppu.OAM[ppu.OAMAddress] = value
         ppu.OAMAddress += 1
     }
@@ -206,7 +206,7 @@ func (ppu *PPUState) CopyOAM(data []byte){
 
     maxOAM := len(ppu.OAM)
     for i := 0; i < len(data); i++ {
-        address := byte(i + ppu.OAMAddress)
+        address := byte(i) + ppu.OAMAddress
         if int(address) >= maxOAM {
             break
         }
